email-sender/requirements.txt
google-cloud-pubsub
functions-framework
kafka-python